% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidygraph.R
\name{minidygraph}
\alias{minidygraph}
\title{Mini version of dygraphs}
\usage{
minidygraph(data, main = NULL, xlab = NULL, ylab = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Time series data,
must be an \link[xts]{xts} object or an object which is convertible to
\code{xts}.}

\item{main}{Main plot title (optional)}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
Interactive dygraph plot
}
\description{
R tutorial of htmlwidgets using \href{http://dygraphs.com}{dygraph} Javascript library.
}
\examples{
library(minidygraphs)
lungDeaths <- cbind(mdeaths, fdeaths)
minidygraph(lungDeaths)
}

